@@ -1,46 +1,83 @@
 {
-    "name": "alpha",
+    "name": "my-bot",
     "version": "1.0.0",
-    "author": "Scott Bucher",
-    "description": "An all-in-one Discord Bot.",
+    "author": "Kevin Novak",
+    "description": "A discord.js bot template written with TypeScript",
     "license": "MIT",
     "private": true,
     "engines": {
-        "node": ">=16.6.0"
+        "node": ">=18.0.0"
     },
-    "main": "dist/start.js",
+    "type": "module",
+    "exports": [
+        "./dist/start-bot.js",
+        "./dist/start-manager.js"
+    ],
     "scripts": {
-        "lint": "tslint --project .",
-        "lint:fix": "tslint --fix --project .",
+        "lint": "eslint . --cache --ext .js,.jsx,.ts,.tsx",
+        "lint:fix": "eslint . --fix --cache --ext .js,.jsx,.ts,.tsx",
         "format": "prettier --check .",
         "format:fix": "prettier --write .",
+        "clean": "git clean -xdf --exclude=\"/config/**/*\"",
+        "clean:dry": "git clean -xdf --exclude=\"/config/**/*\" --dry-run",
         "build": "tsc --project tsconfig.json",
-        "start": "npm run build && node --enable-source-maps dist/start.js",
-        "start:dev": "ts-node-dev --watch src/**/*.ts,config/**/*.json,lang/**/*.json src/start.ts",
-        "start:shard": "npm run build && node --enable-source-maps dist/app.js",
+        "commands:view": "npm run build && node --enable-source-maps dist/start-bot.js commands view",
+        "commands:register": "npm run build && node --enable-source-maps dist/start-bot.js commands register",
+        "commands:rename": "npm run build && node --enable-source-maps dist/start-bot.js commands rename",
+        "commands:delete": "npm run build && node --enable-source-maps dist/start-bot.js commands delete",
+        "commands:clear": "npm run build && node --enable-source-maps dist/start-bot.js commands clear",
+        "start": "npm run start:bot",
+        "start:bot": "npm run build && node --enable-source-maps dist/start-bot.js",
+        "start:manager": "npm run build && node --enable-source-maps dist/start-manager.js",
         "start:pm2": "npm run build && npm run pm2:start",
         "pm2:start": "pm2 start process.json",
         "pm2:stop": "pm2 stop process.json",
-        "pm2:delete": "pm2 delete process.json"
+        "pm2:delete": "pm2 delete process.json",
+        "test": "vitest run",
+        "test:watch": "vitest",
+        "test:coverage": "vitest run --coverage"
     },
     "dependencies": {
-        "discord.js": "13.3.1",
-        "discord.js-collector-utils": "1.2.4",
-        "emoji-regex": "10.0.0",
-        "moment": "2.29.1",
-        "mysql": "2.18.1",
-        "node-fetch": "2.6.5",
-        "node-schedule": "2.0.0",
-        "pm2": "^5.1.2"
+        "@discordjs/rest": "^2.4.3",
+        "@mikro-orm/cli": "^6.1.0",
+        "@mikro-orm/core": "^6.1.0",
+        "@mikro-orm/migrations-mongodb": "^6.1.0",
+        "@mikro-orm/mongodb": "^6.1.0",
+        "@mikro-orm/reflection": "^6.1.0",
+        "@mikro-orm/seeder": "^6.1.0",
+        "class-transformer": "0.5.1",
+        "class-validator": "0.14.1",
+        "cron-parser": "^4.9.0",
+        "discord.js": "^14.18.0",
+        "discord.js-rate-limiter": "1.3.2",
+        "express": "4.21.2",
+        "express-promise-router": "4.1.1",
+        "filesize": "10.1.6",
+        "linguini": "1.3.1",
+        "luxon": "3.5.0",
+        "nanoid": "^5.1.5",
+        "node-fetch": "3.3.2",
+        "node-schedule": "2.1.1",
+        "pino": "9.6.0",
+        "pino-pretty": "13.0.0",
+        "pm2": "^5.4.3",
+        "reflect-metadata": "^0.2.2",
+        "remove-markdown": "0.6.0"
     },
     "devDependencies": {
-        "@types/moment-timezone": "0.5.13",
-        "@types/mysql": "2.15.19",
-        "@types/node": "16.11.6",
-        "@types/node-fetch": "2.5.12",
-        "prettier": "^2.4.1",
-        "ts-node-dev": "^1.1.8",
-        "tslint": "^6.1.3",
-        "typescript": "^4.4.4"
+        "@types/express": "4.17.21",
+        "@types/luxon": "3.4.2",
+        "@types/node": "^22.13.4",
+        "@types/node-schedule": "2.1.7",
+        "@types/remove-markdown": "0.3.4",
+        "@typescript-eslint/eslint-plugin": "^8.24.1",
+        "@typescript-eslint/parser": "^8.24.1",
+        "@vitest/coverage-v8": "^3.0.8",
+        "eslint": "^9.20.1",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-unicorn": "^57.0.0",
+        "prettier": "^3.5.1",
+        "typescript": "^5.7.3",
+        "vitest": "^3.0.8"
     }
 }
