@@ -1,11 +1,31 @@
-import { DMChannel, Message, TextChannel } from 'discord.js';
+import {
+    ApplicationCommandOptionChoiceData,
+    AutocompleteFocusedOption,
+    AutocompleteInteraction,
+    CommandInteraction,
+    PermissionsString,
+} from 'discord.js';
+import { RateLimiter } from 'discord.js-rate-limiter';
+
+import { EventData } from '../models/internal-models.js';
+import { EventDataType } from '../enums/index.js';
 
 export interface Command {
-    name: string;
-    aliases: string[];
-    trigger: RegExp;
-    guildOnly: boolean;
-    adminOnly: boolean;
-    ownerOnly: boolean;
-    execute(args: string[], msg: Message, channel: TextChannel | DMChannel): Promise<void>;
+    names: string[];
+    cooldown?: RateLimiter;
+    deferType: CommandDeferType;
+    requireClientPerms: PermissionsString[];
+    requireEventData: EventDataType[];
+    isDevOnly?: boolean;
+    autocomplete?(
+        intr: AutocompleteInteraction,
+        option: AutocompleteFocusedOption
+    ): Promise<ApplicationCommandOptionChoiceData[]>;
+    execute(intr: CommandInteraction, data: EventData): Promise<void>;
+}
+
+export enum CommandDeferType {
+    PUBLIC = 'PUBLIC',
+    HIDDEN = 'HIDDEN',
+    NONE = 'NONE',
 }
