@@ -1,22 +1,3 @@
-export { AboutCommand } from './about-command';
-export { AddLevelingRewardCommand } from './add-level-reward-command';
-export { AddRoleCallCommand } from './add-role-call-command';
-export { ClearLevelRewardsCommand } from './clear-level-rewards-command';
-export { Command } from './command';
-export { CreateRoleCallCommand } from './create-role-call-command';
-export { DefaultHelpCommand } from './default-help-command';
-export { EightBallCommand } from './eight-ball-command';
-export { GoodMorningCommand } from './good-morning-command';
-export { GoodnightCommand } from './goodnight-command';
-export { PollCommand } from './poll-command';
-export { QuoteCommand } from './quote-command';
-export { RemoveRoleCallCommand } from './remove-role-call-command';
-export { ServerInfoCommand } from './server-info-command';
-export { SetJoinRoleCommand } from './set-join-role-command';
-export { SetLevelingChannelCommand } from './set-leveling-channel-command';
-export { SetPollChannelCommand } from './set-poll-channel-command';
-export { SetQuoteChannelCommand } from './set-quote-channel-command';
-export { SetWelcomeChannelCommand } from './set-welcome-channel-command';
-export { SetXpCommand } from './set-xp-command';
-export { XpCommand } from './xp-command';
-export { XpLeaderboardCommand } from './xp-leaderboard-command';
+export { Args } from './args.js';
+export { Command, CommandDeferType } from './command.js';
+export { ChatCommandMetadata, MessageCommandMetadata, UserCommandMetadata } from './metadata.js';
