@@ -1,5 +1,3 @@
-import { User } from 'discord.js';
-
 export interface EventHandler {
-    process(event: any, user?: User): Promise<void>;
+    process(...args: any[]): Promise<void>;
 }
