@@ -1,9 +1,9 @@
 import fetch, { Response } from 'node-fetch';
-import { URL } from 'url';
+import { URL } from 'node:url';
 
 export class HttpService {
     public async get(url: string | URL, authorization: string): Promise<Response> {
-        return fetch(url, {
+        return await fetch(url.toString(), {
             method: 'get',
             headers: {
                 Authorization: authorization,
@@ -13,26 +13,26 @@ export class HttpService {
     }
 
     public async post(url: string | URL, authorization: string, body?: object): Promise<Response> {
-        return fetch(url, {
+        return await fetch(url.toString(), {
             method: 'post',
             headers: {
                 Authorization: authorization,
                 'Content-Type': 'application/json',
                 Accept: 'application/json',
             },
-            body: body ? JSON.stringify(body) : null,
+            body: body ? JSON.stringify(body) : undefined,
         });
     }
 
     public async put(url: string | URL, authorization: string, body?: object): Promise<Response> {
-        return fetch(url, {
+        return await fetch(url.toString(), {
             method: 'put',
             headers: {
                 Authorization: authorization,
                 'Content-Type': 'application/json',
                 Accept: 'application/json',
             },
-            body: body ? JSON.stringify(body) : null,
+            body: body ? JSON.stringify(body) : undefined,
         });
     }
 
@@ -41,14 +41,14 @@ export class HttpService {
         authorization: string,
         body?: object
     ): Promise<Response> {
-        return fetch(url, {
+        return await fetch(url.toString(), {
             method: 'delete',
             headers: {
                 Authorization: authorization,
                 'Content-Type': 'application/json',
                 Accept: 'application/json',
             },
-            body: body ? JSON.stringify(body) : null,
+            body: body ? JSON.stringify(body) : undefined,
         });
     }
 }
